Refactored project structure
Changed build system from cmake to premake5. Updated .travis.yml,
.gitignore, and readme to reflect these changes. I've been using
premake5 in other projects recently, and it's scripting file is very
simple to generate. This will allow us to add new build rules for other
build targets in the future with ease. linked_list files were also
renamed to linear_linked_list.
