+++ tests/test_config_main.cpp
@@ -1,0 +1,9 @@
+/*
+ * File: test_config_main.cpp
+ *
+ * https://github.com/catchorg/Catch2/blob/master/docs/slow-compiles.md#top
+ */
+
+#define CATCH_CONFIG_MAIN
+#include "catch.hpp"
+
