--- tests/tests_main.cpp
@@ -1,16 +1,0 @@
-/*******************************************************************************
-* File: tests_main.cpp
-*
-* Description: test_main is the entry point for the catch2 unit test framework
-               To reduce compilation time, compile this file first, then link 
-               its object file onto the unit tests source file. 
-*
-* Author: Alexander DuPree
-* 
-* Compiler: GNU GCC 5.4.0
-*
-* Date: 2018 - 05 - 31
-*******************************************************************************/
-
-#define CATCH_CONFIG_MAIN
-#include "catch.hpp"
